run template test commands through jest
