+++ packages/templates/jest.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  collectCoverage: true,
+  preset: "ts-jest",
+  testEnvironment: "node",
+  testMatch: ["**/tests.spec.ts"],
+  globals: {
+    "ts-jest": {
+      diagnostics: false
+    },
+  },
+};
