+++ packages/templates/package.json
@@ -8,13 +8,15 @@
     "url": "https://github.com/polywrap/monorepo.git"
   },
   "scripts": {
-    "test": "ts-node ./run-tests.ts",
+    "test": "jest --passWithNoTests --runInBand --verbose",
     "test:ci": "yarn test"
   },
   "devDependencies": {
-    "ts-node": "8.10.2"
+    "ts-node": "8.10.2",
+    "@types/jest": "26.0.8",
+    "jest": "26.6.3"
   },
-  "gitHead": "7346adaf5adb7e6bbb70d9247583e995650d390a",
+ "gitHead": "7346adaf5adb7e6bbb70d9247583e995650d390a",
   "publishConfig": {
     "access": "public"
   }
